// src/plugin-suggest/types.ts
export enum ChangeReason {
  /**
   * The user has entered or started a new suggestion.
   */
  Start = 'start',

  /**
   * A changed happened to the character. This can be addition, deletion or
   * replacement.
   */
  Text = 'change-character',

  /**
   * A change happened to the selection status which was not purely a move. The
   * selection area may have been increased.
   */
  SelectionInside = 'selection-inside',

  /**
   * The cursor was moved.
   */
  Move = 'move',

  /**
   * The user has moved from one suggestion to another suggestion earlier in the
   * document.
   */
  JumpBackward = 'jump-backward-change',

  /**
   * The user has moved from one suggestion to another suggestion further along
   * in the document.
   */
  JumpForward = 'jump-forward-change'
}

/**
 * The parameters needed for the [[`SuggestIgnoreProps.addIgnored`]] action
 * method available to the suggest plugin handlers.
 *
 * @remarks
 *
 * See:
 * - [[`RemoveIgnoredProps`]]
 */
export interface AddIgnoredProps extends RemoveIgnoredProps {
  /**
   * When `false` this will ignore the range for all matching suggesters. When
   * true the ignored suggesters will only be the ones provided by the name.
   */
  specific?: boolean;
}
// ...
export interface SuggestIgnoreProps {
  /**
   * Add a match target to the ignored matches.
   *
   * @remarks
   *
   * Until the activation character is deleted the `onChange` handler will no
   * longer be triggered for the matched character. It will be like the match
   * doesn't exist.
   *
   * By ignoring the activation character the plugin ensures that any further
   * matches from the activation character will be ignored.
   *
   * There are a number of use cases for this. You may chose to ignore a match
   * when:
   *
   * - The user presses the `escape` key to exit your suggestion dropdown.
   * - The user continues typing without selecting any of the options for the
   *   selection drop down.
   * - The user clicks outside of the suggesters dropdown.
   *
   * ```ts
   * const suggester = {
   *   onChange: ({ addIgnored, range: { from }, suggester: { char, name } }: SuggestExitHandlerProps) => {
   *     addIgnored({ from, char, name }); // Ignore this suggester
   *   },
   * }
   * ```
   */
  addIgnored: (props: AddIgnoredProps) => void;

  /**
   * When the name is provided remove all ignored decorations which match the
   * named suggester. Otherwise remove **all** ignored decorations from the
   * document.
   */
  clearIgnored: (name?: string) => void;

  /**
   * Use this method to skip the next `onChange` when the reason is an exit.
   *
   * @remarks
   *
   * This is useful when you manually call a command which applies the
   * suggestion outside of the `onChange` handler. When that happens `onChange`
   * will still be triggered since the document has now changed. If you don't
   * have the logic set up properly it may rerun your exit logic. This can lead
   * to mismatched transaction errors since the `onChange` handler is provided
   * the last active range and query when the reason is an exit, but these
   * values are probably no longer valid.
   *
   * This helper method can be applied to make life easier. Call it when running
   * a command in a click handler or key binding and you don't have to worry
   * about your `onChange` handler being called again and leading to a
   * mismatched transaction error.
   */
  ignoreNextExit: () => void;
}
// ...
export interface SuggestMatch extends SuggesterProps {
  /**
   * Range of current match; for example `@foo|bar` (where | is the cursor)
   * - `from` is the start (= 0)
   * - `to` is cursor position (= 4)
   * - `end` is the end of the match (= 7)
   */
  range: RangeWithCursor;

  /**
   * Current query of match which doesn't include the activation character.
   */
  query: MatchValue;

  /**
   * Full text of match including the activation character
   *
   * @remarks
   *
   * For a `char` of `'@'` and query of `'awesome'` `text.full` would be
   * `'@awesome'`.
   */
  text: MatchValue;

  /**
   * The raw regex match which caused this suggester to be triggered.
   *
   * - `rawMatch[0]` is always the full match.
   * - `rawMatch[1]` is always the matching character (or regex pattern).
   *
   * To make use of this you can set the [[`Suggester.supportedCharacters`]]
   * property to be a regex which included matching capture group segments.
   */
  match: RegExpExecArray;

  /**
   * The text after full match, up until the end of the text block.
   */
  textAfter: string;

  /**
   * The text before the full match, up until the beginning of the node.
   */
  textBefore: string;
}
// ...
export interface SuggestMarkProps {
  /**
   * When managing suggesters with marks it is possible to remove a mark without
   * the change reflecting in the prosemirror state. This method should be used
   * when removing a suggestion if you are using prosemirror `Marks` to identify
   * the suggestion.
   *
   * When this method is called, `prosemirror-suggest` will handle the removal
   * of the mark in the next state update (during apply).
   */
  setMarkRemoved: () => void;
}

/**
 * A parameter builder interface indicating the reason the handler was called.
 *
 * @typeParam Reason - Whether this is change or an exit reason.
 */
export interface ReasonProps {
  /**
   * The reason for the exit. Either this or the change reason must have a
   * value.
   */
  exitReason?: ExitReason;

  /**
   * The reason for the change. Either this or change reason must have a value..
   */
  changeReason?: ChangeReason;
}

/**
 * The parameter passed to the  [[`Suggester.onChange`]] method. It the
 * properties `changeReason` and `exitReason` which are available depending on
 * whether this is an exit or change.
 *
 * Exactly **ONE** will always be available. Unfortunately that's quite hard to
 * model in TypeScript without complicating all dependent types.
 */
export interface SuggestChangeHandlerProps
  extends SuggestMatchWithReason,
    EditorViewProps,
    SuggestIgnoreProps,
    SuggestMarkProps,
    Pick<Suggester, 'name' | 'char'> {}
// ...
export interface SuggestMatchWithReason extends SuggestMatch, ReasonProps {}
// ...
export type EditorView = PMView.EditorView;

/**
 * A parameter builder interface containing the `view` property.
 *
 * @typeParam Schema - the underlying editor schema.
 */
export interface EditorViewProps {
  /**
   * An instance of the ProseMirror editor `view`.
   */
  view: EditorView;
}